pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["Yasaman karimi <yasmin.kmi.ag@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.3"
django-ninja = "^1.3.0"
docusign-esign = {git = "https://github.com/docusign/docusign-esign-python-client.git"}
pyjwt = "^2.10.1"
pip = "^24.3.1"
cryptography = "^44.0.0"
django-shell-plus = "^1.1.7"
django-environ = "^0.11.2"
django-cors-headers = "^4.6.0"
redis = "^5.2.0"
pydantic = "^2.10.2"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
